+++ exporters/trace/jaeger/doc.go
@@ -14,7 +14,8 @@
 
 // Package jaeger contains an OpenTelemetry tracing exporter for Jaeger.
 //
-// This package is currently in a pre-GA phase. Backwards incompatible changes
-// may be introduced in subsequent minor version releases as we work to track
-// the evolving OpenTelemetry specification and user feedback.
+// Deprecated: Use Jaeger exporter in package
+// go.opentelemetry.io/otel/exporters/jaeger.
+//
+// This package is frozen and no new functionality will be added.
 package jaeger // import "go.opentelemetry.io/otel/exporters/trace/jaeger"